Clean wallpost photo api
