+++ apps/projects/urlsapi.py
@@ -1,9 +1,7 @@
 from apps.projects.views import ManageProjectAmbassadorList, ManageProjectAmbassadorDetail, ManageProjectBudgetLinetList, ManageProjectBudgetLineDetail, ProjectPreviewList, ProjectPreviewDetail
 from django.conf.urls import patterns, url, include
 from surlex.dj import surl
-from .views import (ProjectDetail, ProjectList, ProjectWallPostList, ProjectWallPostDetail, ProjectMediaWallPostList,
-                    ProjectMediaWallPostDetail, ProjectTextWallPostList, ProjectTextWallPostDetail,
-                    ProjectMediaWallPostPhotoList, ProjectMediaWallPostPhotoDetail, ProjectSupporterList,
+from .views import (ProjectDetail, ProjectList, ProjectSupporterList,
                     ManageProjectList, ManageProjectDetail, ManageProjectPitchDetail, ManageProjectPlanDetail,
                     ProjectPitchDetail, ProjectPlanDetail, ManageProjectCampaignDetail, ProjectThemeList,
                     ProjectThemeDetail, ProjectDonationList)
@@ -18,18 +16,6 @@ urlpatterns = patterns('',
     # Not publically avaialable atm
     # surl(r'^pitches/<pk:#>$', ProjectPitchDetail.as_view(), name='project-pitch-detail'),
     surl(r'^plans/<pk:#>$', ProjectPlanDetail.as_view(), name='project-plan-detail'),
-
-
-    # Project WallPost Urls
-    url(r'^wallposts/$', ProjectWallPostList.as_view(), name='project-wallpost-list'),
-    surl(r'^wallposts/<pk:#>$', ProjectWallPostDetail.as_view(), name='project-wallpost-detail'),
-    url(r'^wallposts/media/$', ProjectMediaWallPostList.as_view(), name='project-mediawallpost-list'),
-    surl(r'^wallposts/media/<pk:#>$', ProjectMediaWallPostDetail.as_view(), name='project-mediawallpost-detail'),
-    url(r'^wallposts/text/$', ProjectTextWallPostList.as_view(), name='project-textwallpost-list'),
-    surl(r'wallposts/text/<pk:#>$', ProjectTextWallPostDetail.as_view(), name='project-textwallpost-detail'),
-
-    url(r'^wallposts/media/photos/$', ProjectMediaWallPostPhotoList.as_view(), name='project-mediawallpost-photo-list'),
-    surl(r'^wallposts/media/photos/<pk:#>$', ProjectMediaWallPostPhotoDetail.as_view(), name='project-mediawallpost-photo-list'),
 
     surl(r'^themes/$', ProjectThemeList.as_view(), name='project-theme-list'),
     surl(r'^themes/<pk:#>$', ProjectThemeDetail.as_view(), name='project-theme-detail'),
