+++ apps/tasks/templates/tasks.hbs
@@ -527,43 +527,35 @@
 {% tplhandlebars "projectTask/index" %}
     {{!-- Project Wall Post List template --}}
 
-    <div id="l-section">
-        <section class="l-wrapper">
-        	<div class="l-content" id="wallposts">
+    {{#if controllers.currentUser.isAuthenticated }}
+        {{#if controllers.projectTask.model.isAuthor}}
+            <h3>{% trans "Write an update" %}</h3>
+            <p>{% trans "Write an update of your task to inform your crowd." %}</p>
+        {{else}}
+            <h3>{% trans "Write a comment" %}</h3>
+            <p>{% trans "Write an comment to place on the task's wall." %}</p>
+        {{/if}}
+        {{ render "taskTextWallPostNew" }}
+    {{else}}
+        <h3>{% trans "Write a comment" %}</h3>
+        <a {{action openInBox "login"}}>{%  trans "Login" %}</a> or become a {{#linkTo "signup"}}{% trans "member" %}{{/linkTo}} {% trans "to leave a comment" %}.</p>
+    {{/if}}
 
-                {{#if controllers.currentUser.isAuthenticated }}
-                    {{#if controllers.projectTask.model.isAuthor}}
-                        <h3>{% trans "Write an update" %}</h3>
-                        <p>{% trans "Write an update of your task to inform your crowd." %}</p>
-                    {{else}}
-                        <h3>{% trans "Write a comment" %}</h3>
-                        <p>{% trans "Write an comment to place on the task's wall." %}</p>
-                    {{/if}}
-                    {{ render "taskTextWallPostNew" }}
-                {{else}}
-                    <h3>{% trans "Write a comment" %}</h3>
-                    <a {{action openInBox "login"}}>{%  trans "Login" %}</a> or become a {{#linkTo "signup"}}{% trans "member" %}{{/linkTo}} {% trans "to leave a comment" %}.</p>
-                {{/if}}
+    {{#each item in controller itemController="wallPost"}}
+        {{#if item.isLoaded}}
+            {{#if item.isSystemWallPost}}
+                {{view "App.SystemWallPostView"}}
+            {{else}}
+                {{view "App.MediaWallPostView"}}
+            {{/if}}
+        {{/if}}
+    {{/each}}
 
-    	        {{#each item in controller itemController="wallPost"}}
-    	            {{#if item.isLoaded}}
-                        {{#if item.isSystemWallPost}}
-        	                {{view "App.SystemWallPostView"}}
-                        {{else}}
-        	                {{view "App.MediaWallPostView"}}
-                        {{/if}}
-    	            {{/if}}
-    	        {{/each}}
-
-                {{#if canLoadMore}}
-                    <a {{action "showMore"}} class="btn-link btn-more">
-                        <span class="flaticon solid plus-2"></span> {% trans "Show more" %} <small>{{remainingItemCount}} {% trans "more" %}</small>
-                    </a>
-                {{/if}}
-
-        	</div>
-        </section>
-    </div>
+    {{#if canLoadMore}}
+        <a {{action "showMore"}} class="btn-link btn-more">
+            <span class="flaticon solid plus-2"></span> {% trans "Show more" %} <small>{{remainingItemCount}} {% trans "more" %}</small>
+        </a>
+    {{/if}}
 
 {% endtplhandlebars %}
